fixed return value of function to build under ubuntu20 on RPi.

+++ utils/pinprocfw/pinprocfw.cpp
@@ -2140,7 +2140,7 @@ int processFile()
             return iErrorCode;
 }
 
-int checkPROCFile() {
+int64_t checkPROCFile() {
     uint32_t checksum=0, file_checksum, file_board_id, header_checksum;
     unsigned char data;
     int i=0,file_i=0;
